+++ kernel/src/lib.rs
@@ -7,9 +7,9 @@
 //! and reuse. The binary entry point is in main.rs.
 
 #![no_std]
-#![feature(custom_test_frameworks)]
-#![test_runner(crate::test_runner)]
-#![reexport_test_harness_main = "test_main"]
+// #![feature(custom_test_frameworks)]
+// #![test_runner(crate::test_runner)]
+// #![reexport_test_harness_main = "test_main"]
 
 extern crate alloc;